+++ torch/_library/__init__.py
@@ -1,0 +1,1 @@
+import torch._library.abstract_impl
